pyproject.toml
[project]
name = "serverless-genai-log-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"   
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.1",
    # GenAI & Agents
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    # PII & Security (Phase 6)
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "spacy>=3.7.0",
    # Storage & Vector (Phase 6 Ready)
    "chromadb>=0.5.0",
    "deltalake>=0.20.0",
    "pandas>=2.2.0",
    "pyarrow>=17.0.0",
    # Ingestion (Phase 6 Ready)
    "aiokafka>=0.11.0",
    # Utilities
    "markdown>=3.7",
    "pyyaml>=6.0",
]
